refactor(consensus): split lib.rs into per-concern modules
tsoracle-consensus carried four distinct concerns in one lib.rs: the
ConsensusDriver trait, ConsensusError, LeaderState, and the AdvancePayload
command plus its range guard. Split them into focused modules and reduce
lib.rs to module declarations and re-exports:

- advance.rs    — AdvancePayload, AdvanceOutOfRange, reject_out_of_range_advance
- error.rs      — ConsensusError
- leadership.rs — LeaderState
- driver.rs     — the ConsensusDriver trait

Each test moves next to the item it covers. The public API is unchanged:
the crate root re-exports every type and the free function, so dependents
keep importing `tsoracle_consensus::{ConsensusError, LeaderState, ...}`
exactly as before. No behavior change.
